backend/requirements.txt
django==4.1
djangorestframework
djoser
django-cors-headers
djangorestframework-simplejwt

